+++ .env.example 
@@ -1,0 +1,3 @@
+ACCESS_TOKEN=YOUR TOKEN
+SERVER_HOSTNAME= YOUR HOSTNAME
+HTTP_PATH=YOUR PATH
