+++ .gitignore
@@ -1,2 +1,3 @@
 cache/
 *.pyc
+.env
